+++ lib/calendar_day_view.dart
@@ -4,6 +4,7 @@ export 'src/day_views/calendar_day_view_base.dart';
 export 'src/day_views/category/category_2d_day_view.dart';
 export 'src/day_views/category/category_day_view_controller.dart';
 export 'src/day_views/day_view_config.dart';
+export 'src/day_views/day_view_decoration.dart';
 export 'src/day_views/multi_column/multi_column_calendar_day_view.dart';
 export 'src/day_views/overflow/overflow_calendar_day_view.dart';
 export 'src/models/categorized_day_event.dart';