+++ lib/src/day_views/in_row_calendar_day_view.dart
@@ -167,11 +167,11 @@ class InRowEventRowWidget<T extends Object> extends StatelessWidget {
       ),
       child: Stack(
         children: [
-          if (config.timeRowBackgroundBuilder != null)
+          if (config.decoration.rowBackground != null)
             Positioned.fill(
               child: Builder(
                 builder: (context) {
-                  final bg = config.timeRowBackgroundBuilder!(
+                  final bg = config.decoration.rowBackground!(
                     context,
                     time,
                     BoxConstraints.tightFor(width: viewWidth, height: config.rowHeight),
@@ -180,16 +180,16 @@ class InRowEventRowWidget<T extends Object> extends StatelessWidget {
                 },
               ),
             ),
-          if (config.dividerBuilder != null)
+          if (config.decoration.divider != null)
             Builder(
-              builder: (context) => config.dividerBuilder!(context, time) ?? const SizedBox.shrink(),
+              builder: (context) => config.decoration.divider!(context, time) ?? const SizedBox.shrink(),
             )
           else
             Divider(
-              color: config.dividerColor ?? Colors.amber,
+              color: config.decoration.dividerColor ?? Colors.amber,
               height: 0,
               thickness: time.minute == 0 ? 1 : .5,
-              indent: config.timeColumnWidth + 3,
+              indent: config.decoration.timeColumnWidth + 3,
             ),
           Row(
             mainAxisSize: MainAxisSize.max,
@@ -199,13 +199,13 @@ class InRowEventRowWidget<T extends Object> extends StatelessWidget {
                 transform: Matrix4.translationValues(0, -20, 0),
                 child: SizedBox(
                   height: 40,
-                  width: config.timeColumnWidth,
-                  child: config.timeLabelBuilder?.call(context, time) ??
+                  width: config.decoration.timeColumnWidth,
+                  child: config.decoration.timeLabel?.call(context, time) ??
                       FittedBox(
                         fit: BoxFit.scaleDown,
                         child: Text(
                           config.time12 ? time.hourDisplay12 : time.hourDisplay24,
-                          style: config.timeTextStyle,
+                          style: config.decoration.timeTextStyle,
                           maxLines: 1,
                         ),
                       ),
@@ -261,12 +261,12 @@ class InRowEventRowWidget<T extends Object> extends StatelessWidget {
 
   Widget _buildCurrentTimeLine() {
     final top = (currentTime.minute - time.minute) * config.heightPerMin;
-    if (config.currentTimeLineBuilder != null) {
-      return config.currentTimeLineBuilder!(top, viewWidth);
+    if (config.decoration.currentTimeLine != null) {
+      return config.decoration.currentTimeLine!(top, viewWidth);
     }
     return CurrentTimeLineWidget(
       top: top,
-      color: config.currentTimeLineColor,
+      color: config.decoration.currentTimeLineColor,
       width: viewWidth,
     );
   }