+++ lib/src/day_views/overflow/overflow_calendar_day_view.dart
@@ -117,7 +117,7 @@ class _OverFlowCalendarDayViewState<T extends Object> extends State<OverFlowCale
     final totalHeight = widget.config.timeList.length * widget.config.rowHeight;
     final viewWidth = MediaQuery.sizeOf(context).width;
 
-    final eventColumnWith = viewWidth - widget.config.timeColumnWidth;
+    final eventColumnWith = viewWidth - widget.config.decoration.timeColumnWidth;
 
     return SafeArea(
       child: SingleChildScrollView(
@@ -140,7 +140,7 @@ class _OverFlowCalendarDayViewState<T extends Object> extends State<OverFlowCale
                     viewWidth: viewWidth,
                     config: widget.config,
                     onTimeTap: widget.onTimeTap,
-                    timeLabelBuilder: widget.config.timeLabelBuilder,
+                    timeLabelBuilder: widget.config.decoration.timeLabel,
                   );
                 },
               ),
@@ -156,12 +156,12 @@ class _OverFlowCalendarDayViewState<T extends Object> extends State<OverFlowCale
                         cropBottomEvents: widget.config.cropBottomEvents,
                         timeStart: widget.config.timeStart,
                         totalHeight: totalHeight,
-                        timeTitleColumnWidth: widget.config.timeColumnWidth,
+                        timeTitleColumnWidth: widget.config.decoration.timeColumnWidth,
                       )
                     : OverflowFixedWidthEventsWidget(
                         heightUnit: widget.config.heightPerMin,
                         eventColumnWidth: eventColumnWith,
-                        timeTitleColumnWidth: widget.config.timeColumnWidth,
+                        timeTitleColumnWidth: widget.config.decoration.timeColumnWidth,
                         timeStart: widget.config.timeStart,
                         overflowEvents: _overflowEvents,
                         overflowItemBuilder: widget.overflowItemBuilder!,
@@ -180,13 +180,13 @@ class _OverFlowCalendarDayViewState<T extends Object> extends State<OverFlowCale
 
   Widget _buildCurrentTimeLine(double viewWidth) {
     final top = _currentTime.minuteFrom(widget.config.timeStart).toDouble() * widget.config.heightPerMin;
-    if (widget.config.currentTimeLineBuilder != null) {
-      return widget.config.currentTimeLineBuilder!(top, viewWidth);
+    if (widget.config.decoration.currentTimeLine != null) {
+      return widget.config.decoration.currentTimeLine!(top, viewWidth);
     }
     return CurrentTimeLineWidget(
       top: top,
       width: viewWidth,
-      color: widget.config.currentTimeLineColor,
+      color: widget.config.decoration.currentTimeLineColor,
     );
   }
 }
@@ -231,11 +231,11 @@ class OverflowTimeRowWidget extends StatelessWidget {
         width: viewWidth,
         child: Stack(
           children: [
-            if (config.timeRowBackgroundBuilder != null)
+            if (config.decoration.rowBackground != null)
               Positioned.fill(
                 child: Builder(
                   builder: (context) {
-                    final bg = config.timeRowBackgroundBuilder!(
+                    final bg = config.decoration.rowBackground!(
                       context,
                       time,
                       BoxConstraints.tightFor(width: viewWidth, height: config.rowHeight),
@@ -244,29 +244,29 @@ class OverflowTimeRowWidget extends StatelessWidget {
                   },
                 ),
               ),
-            if (config.dividerBuilder != null)
+            if (config.decoration.divider != null)
               Builder(
-                builder: (context) => config.dividerBuilder!(context, time) ?? const SizedBox.shrink(),
+                builder: (context) => config.decoration.divider!(context, time) ?? const SizedBox.shrink(),
               )
             else
               Divider(
-                color: config.dividerColor ?? Colors.amber,
+                color: config.decoration.dividerColor ?? Colors.amber,
                 height: 0,
                 thickness: time.minute == 0 ? 1 : .5,
-                indent: config.timeColumnWidth + 3,
+                indent: config.decoration.timeColumnWidth + 3,
               ),
             Transform(
               transform: Matrix4.translationValues(0, -20, 0),
               child: SizedBox(
                 height: 40,
-                width: config.timeColumnWidth,
+                width: config.decoration.timeColumnWidth,
                 child: Center(
                   child: timeLabelBuilder?.call(context, time) ??
                       FittedBox(
                         fit: BoxFit.scaleDown,
                         child: Text(
                           config.time12 ? time.hourDisplay12 : time.hourDisplay24,
-                          style: config.timeTextStyle ?? TextStyle(color: config.timeTextColor),
+                          style: config.decoration.timeTextStyle ?? TextStyle(color: config.decoration.timeTextColor),
                           maxLines: 1,
                         ),
                       ),